MOD: ACAMP version 2.0 & add a large of comments
