+++ AcampProtocol.c
@@ -1,0 +1,278 @@
+#include "AcampProtocol.h"
+
+/**
+ * APProtocolStore8
+ * store a 8-bit value into a initialized APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a initialized APProtocolMessage and there is space left
+ * @param u8 val, value you want to store
+ */
+void APProtocolStore8(APProtocolMessage *msgPtr, u8 val)
+{
+	AP_COPY_MEMORY(&((msgPtr->msg)[(msgPtr->offset)]), &(val), 1);
+	(msgPtr->offset) += 1;
+}
+
+/**
+ * APProtocolStore16
+ * store a 16-bit value into a initialized APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a initialized APProtocolMessage and there is space left
+ * @param u16 val, value you want to store
+ */
+void APProtocolStore16(APProtocolMessage *msgPtr, u16 val)
+{
+	val = htons(val);
+	AP_COPY_MEMORY(&((msgPtr->msg)[(msgPtr->offset)]), &(val), 2);
+	(msgPtr->offset) += 2;
+}
+
+/**
+ * APProtocolStore32
+ * store a 32-bit value into a initialized APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a initialized APProtocolMessage and there is space left
+ * @param u32 val, value you want to store
+ */
+void APProtocolStore32(APProtocolMessage *msgPtr, u32 val)
+{
+	val = htonl(val);
+	AP_COPY_MEMORY(&((msgPtr->msg)[(msgPtr->offset)]), &(val), 4);
+	(msgPtr->offset) += 4;
+}
+
+/**
+ * APProtocolStore1Str
+ * store a string into a initialized APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a initialized APProtocolMessage and there is space left
+ * @param char *str, value you want to store, a standard string must end with '\0'
+ */
+void APProtocolStoreStr(APProtocolMessage *msgPtr, char *str)
+{
+	int len = strlen(str);
+	AP_COPY_MEMORY(&((msgPtr->msg)[(msgPtr->offset)]), str, len);
+	(msgPtr->offset) += len;
+}
+
+/**
+ * APProtocolStore1Message
+ * store content in another APProtocolMessage into a initialized APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a initialized APProtocolMessage and there is space left
+ * @param APProtocolMessage *msgToStorePtr, another APProtocolMessage you want to use
+ */
+void APProtocolStoreMessage(APProtocolMessage *msgPtr, APProtocolMessage *msgToStorePtr)
+{
+	AP_COPY_MEMORY(&((msgPtr->msg)[(msgPtr->offset)]), msgToStorePtr->msg, msgToStorePtr->offset);
+	(msgPtr->offset) += msgToStorePtr->offset;
+}
+
+/**
+ * APProtocolStore1RawBytes
+ * store raw bytes into a initialized APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a initialized APProtocolMessage and there is space left
+ * @param u8 *bytes, value you want to store
+ * @param int len, size of raw bytes
+ */
+void APProtocolStoreRawBytes(APProtocolMessage *msgPtr, u8 *bytes, int len)
+{
+	AP_COPY_MEMORY(&((msgPtr->msg)[(msgPtr->offset)]), bytes, len);
+	(msgPtr->offset) += len;
+}
+
+/**
+ * APProtocolStore1Reserved
+ * store reserved value into a initialized APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a initialized APProtocolMessage and there is space left
+ * @param int len, size of reserved space
+ */
+void APProtocolStoreReserved(APProtocolMessage *msgPtr, int reservedLen)
+{
+	u8* reserved;
+	AP_CREATE_OBJECT_SIZE(reserved, reservedLen);
+	AP_ZERO_MEMORY(reserved, reservedLen);
+	APProtocolStoreRawBytes(msgPtr, reserved, reservedLen);
+	AP_FREE_OBJECT(reserved);
+}
+
+/**
+ * APAssembleMsgElem
+ * assemble a formatted msg element by the element content and type
+ * @param APProtocolMessage *msgPtr, the element content, which is preserved in a unformatted APProtocolMessage. after this action, it will transform a formatted msg element
+ * @param u16 type, the type of msg elem
+ * @return APBool, whether the operation is success or not
+ */
+APBool APAssembleMsgElem(APProtocolMessage *msgPtr, u16 type)
+{
+	APProtocolMessage completeMsg;
+
+	if(msgPtr == NULL) return AP_FALSE;
+	AP_INIT_PROTOCOL_MESSAGE(completeMsg, ELEMENT_HEADER_LEN+(msgPtr->offset));
+
+	APProtocolStore16(&completeMsg, type);
+	APProtocolStore16(&completeMsg, msgPtr->offset);
+	APProtocolStoreMessage(&completeMsg, msgPtr);
+
+	AP_FREE_PROTOCOL_MESSAGE(*msgPtr);
+
+	msgPtr->msg = completeMsg.msg;
+	msgPtr->offset = completeMsg.offset;
+	msgPtr->type = type;
+
+	return AP_TRUE;
+}
+
+/**
+ * APAssembleControlMessage
+ * assemble a formatted msg that can be send
+ * @param APProtocolMessage *msgPtr, a APAssembleMessage type value, which will output a formatted msg
+ * @param u16 apid, apid that will be used in message header
+ * @param u32 seqNum, seq num that will be used in message header
+ * @param u16 msgType, the type of msg
+ * @param APProtocolMessage *msgElems, a array of msgElems, which will be used in msg
+ * @param int msgElemNum, count of msgElems
+ * @return APBool, whether the operation is success or not
+ */
+APBool APAssembleControlMessage(APProtocolMessage *msgPtr, u16 apid, u32 seqNum,
+						 u16 msgType, APProtocolMessage *msgElems, int msgElemNum)
+{
+	APProtocolMessage controlHdr, completeMsg;
+	int msgElemsLen = 0, i;
+	APHeaderVal controlHdrVal;
+
+	if(msgElems == NULL && msgElemNum > 0)
+		return AP_FALSE;
+
+	for(i = 0; i < msgElemNum; i++) msgElemsLen += msgElems[i].offset;
+
+	controlHdrVal.version = CURRENT_VERSION;
+	controlHdrVal.type = TYPE_CONTROL;
+	controlHdrVal.apid = apid;
+	controlHdrVal.seqNum = seqNum;
+	controlHdrVal.msgType = msgType;
+	controlHdrVal.msgLen = HEADER_LEN + msgElemsLen;
+
+	if(!(APAssembleControlHeader(&controlHdr, &controlHdrVal))) {
+		AP_FREE_PROTOCOL_MESSAGE(controlHdr);
+		AP_FREE_ARRAY_AND_PROTOCOL_MESSAGE(msgElems, msgElemNum);
+		return AP_FALSE;
+	}
+
+	AP_INIT_PROTOCOL_MESSAGE(completeMsg, controlHdr.offset + msgElemsLen);
+	APProtocolStoreMessage(&completeMsg, &controlHdr);
+	AP_FREE_PROTOCOL_MESSAGE(controlHdr);
+
+	for(i = 0; i < msgElemNum; i++) {
+		APProtocolStoreMessage(&completeMsg, &(msgElems[i]));
+	}
+	AP_FREE_ARRAY_AND_PROTOCOL_MESSAGE(msgElems, msgElemNum);
+
+	AP_FREE_PROTOCOL_MESSAGE(*msgPtr);
+	msgPtr->msg = completeMsg.msg;
+	msgPtr->offset = completeMsg.offset;
+	msgPtr->type = msgType;
+
+	return AP_TRUE;
+}
+
+/**
+ * APAssembleControlHeader
+ * assemble a msg header
+ * @param APProtocolMessage *controlHdrPtr, a APAssembleMessage type value, which will output a formatted msg header
+ * @param APHeaderVal *valPtr, header info
+ * @return APBool, whether the operation is success or not
+ */
+APBool APAssembleControlHeader(APProtocolMessage *controlHdrPtr, APHeaderVal *valPtr)
+{
+	if(controlHdrPtr == NULL || valPtr == NULL) return AP_FALSE;
+
+	AP_INIT_PROTOCOL_MESSAGE(*controlHdrPtr, HEADER_LEN);
+
+	APProtocolStore8(controlHdrPtr, valPtr->version);
+	APProtocolStore8(controlHdrPtr, valPtr->type);
+	APProtocolStore16(controlHdrPtr, valPtr->apid);
+	APProtocolStore32(controlHdrPtr, valPtr->seqNum);
+	APProtocolStore16(controlHdrPtr, valPtr->msgType);
+	APProtocolStore16(controlHdrPtr, valPtr->msgLen);
+	APProtocolStoreReserved(controlHdrPtr, 4);
+
+	return AP_TRUE;
+}
+
+/**
+ * APProtocolRetrieve8
+ * get a 8-bit value from APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a APProtocolMessage that include value
+ * @return u8, u8 value
+ */
+u8 APProtocolRetrieve8(APProtocolMessage *msgPtr) {
+	u8 val;
+
+	AP_COPY_MEMORY(&val, &((msgPtr->msg)[(msgPtr->offset)]), 1);
+	(msgPtr->offset) += 1;
+
+	return val;
+}
+
+/**
+ * APProtocolRetrieve16
+ * get a 16-bit value from APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a APProtocolMessage that include value
+ * @return u16, u16 value
+ */
+u16 APProtocolRetrieve16(APProtocolMessage *msgPtr) {
+	u16 val;
+
+	AP_COPY_MEMORY(&val, &((msgPtr->msg)[(msgPtr->offset)]), 2);
+	(msgPtr->offset) += 2;
+
+	return ntohs(val);
+}
+
+/**
+ * APProtocolRetrieve32
+ * get a 32-bit value from APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a APProtocolMessage that include value
+ * @return u32, u32 value
+ */
+u32 APProtocolRetrieve32(APProtocolMessage *msgPtr) {
+	u32 val;
+
+	AP_COPY_MEMORY(&val, &((msgPtr->msg)[(msgPtr->offset)]), 4);
+	(msgPtr->offset) += 4;
+
+	return ntohl(val);
+}
+
+/**
+ * APProtocolRetrieveStr
+ * get a string from APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a APProtocolMessage that include value
+ * @param int len, length of string
+ * @return char*, a standard string that end with '\0'
+ */
+char* APProtocolRetrieveStr(APProtocolMessage *msgPtr, int len) {
+	u8* str;
+
+	AP_CREATE_OBJECT_SIZE(str, (len+1));
+
+	AP_COPY_MEMORY(str, &((msgPtr->msg)[(msgPtr->offset)]), len);
+	str[len] = '\0';
+	(msgPtr->offset) += len;
+
+	return (char*)str;
+}
+
+/**
+ * APProtocolRetrieveRawBytes
+ * get raw bytes from APProtocolMessage
+ * @param APProtocolMessage *msgPtr, a APProtocolMessage that include value
+ * @param int len, size of raw bytes
+ * @return u8*, an array that include raw bytes
+ */
+u8* APProtocolRetrieveRawBytes(APProtocolMessage *msgPtr, int len) {
+	u8* bytes;
+
+	AP_CREATE_OBJECT_SIZE(bytes, len);
+
+	AP_COPY_MEMORY(bytes, &((msgPtr->msg)[(msgPtr->offset)]), len);
+	(msgPtr->offset) += len;
+
+	return bytes;
+}
