+++ Common.h
@@ -28,8 +28,8 @@ typedef enum  {
 
 #define	AP_CREATE_OBJECT_SIZE(obj_name, obj_size)	obj_name = ((u8*)malloc(obj_size));
 #define	AP_CREATE_STRING_SIZE(str_name, str_size)	str_name = ((char*)malloc(str_size));
-#define	AP_FREE_OBJECT(obj_name)		{if(obj_name){free((obj_name)); (obj_name) = NULL;}}
-#define	AP_FREE_OBJECTS_ARRAY(ar_name, ar_size)	{int _i = 0; for(_i = ((ar_size)-1); _i >= 0; _i--) {if(((ar_name)[_i]) != NULL){ free((ar_name)[_i]);}} free(ar_name); (ar_name) = NULL; }
+#define	AP_FREE_OBJECT(obj_name)					{if(obj_name){free((obj_name)); (obj_name) = NULL;}}
+#define	AP_FREE_OBJECTS_ARRAY(ar_name, ar_size)		{int _i = 0; for(_i = ((ar_size)-1); _i >= 0; _i--) {if(((ar_name)[_i]) != NULL){ free((ar_name)[_i]);}} free(ar_name); (ar_name) = NULL; }
 
 #define	AP_CREATE_ARRAY(ar_name, ar_size, ar_type)	{ar_name = (ar_type*) (malloc(sizeof(ar_type) * (ar_size)));}
